// client/src/app/stores/products/product.types.ts
export enum PRODUCTS_ACTION_TYPES {
  // Fetch Products types
  FETCH_PRODUCTS_LOADING = "products/FETCH_PRODUCTS_LOADING",
  FETCH_PRODUCTS_SUCCESS = "products/FETCH_PRODUCTS_SUCCESS",
  FETCH_PRODUCTS_FAILED = "products/FETCH_PRODUCTS_FAILED",
  // Fetch Product for admin types
  FETCH_PRODUCT_ADMIN_LOADING = "products/FETCH_PRODUCT_ADMIN_LOADING",
  FETCH_PRODUCT_ADMIN_SUCCESS = "products/FETCH_PRODUCT_ADMIN_SUCCESS",
  FETCH_PRODUCT_ADMIN_FAILED = "products/FETCH_PRODUCT_ADMIN_FAILED",
  // Fetch Product by name types
  FETCH_PRODUCT_BY_NAME_LOADING = "products/FETCH_PRODUCT_BY_NAME_LOADING",
  FETCH_PRODUCT_BY_NAME_SUCCESS = "products/FETCH_PRODUCT_BY_NAME_SUCCESS",
  FETCH_PRODUCT_BY_NAME_FAILED = "products/FETCH_PRODUCT_BY_NAME_FAILED",
  // Fetch Products for admin types
  FETCH_PRODUCTS_ADMIN_LOADING = "products/FETCH_PRODUCTS_ADMIN_LOADING",
  FETCH_PRODUCTS_ADMIN_SUCCESS = "products/FETCH_PRODUCTS_ADMIN_SUCCESS",
  FETCH_PRODUCTS_ADMIN_FAILED = "products/FETCH_PRODUCTS_ADMIN_FAILED",
  // Fetch Products by vendor types
  FETCH_PRODUCTS_BY_VENDOR_LOADING = "products/FETCH_PRODUCTS_BY_VENDOR_LOADING",
  FETCH_PRODUCTS_BY_VENDOR_SUCCESS = "products/FETCH_PRODUCTS_BY_VENDOR_SUCCESS",
  FETCH_PRODUCTS_BY_VENDOR_FAILED = "products/FETCH_PRODUCTS_VENDOR_FAILED",
  // Add product types
  ADD_PRODUCT_LOADING = "products/ADD_PRODUCT_LOADING",
  ADD_PRODUCT_SUCCESS = "products/ADD_PRODUCT_SUCCESS",
  ADD_PRODUCT_FAILED = "products/ADD_PRODUCT_FAILED",
  // Update Product types
  UPDATE_PRODUCT_LOADING = "products/UPDATE_PRODUCT_LOADING",
  UPDATE_PRODUCT_SUCCESS = "products/UPDATE_PRODUCT_SUCCESS",
  UPDATE_PRODUCT_FAILED = "products/UPDATE_PRODUCT_FAILED",
  // Delete Product types
  DELETE_PRODUCT_LOADING = "products/DELETE_PRODUCT_LOADING",
  DELETE_PRODUCT_SUCCESS = "products/DELETE_PRODUCT_SUCCESS",
  DELETE_PRODUCT_FAILED = "products/DELETE_PRODUCT_FAILED",
  // Search Product types
  SEARCH_PRODUCTS_BY_TERM_LOADING = "products/SEARCH_PRODUCTS_BY_TERM_LOADING",
  SEARCH_PRODUCTS_BY_TERM_SUCCESS = "products/SEARCH_PRODUCTS_BY_TERM_SUCCESS",
  SEARCH_PRODUCTS_BY_TERM_FAILED = "products/SEARCH_PRODUCTS_BY_TERM_FAILED",
};